+++ Kernel/Modules/AdminPackageManager.pm
@@ -1282,6 +1282,34 @@ sub Run {
                 %Errors,
             },
         );
+
+        # check if we're on MySQL and show a max_allowed_packet notice
+        # if the actual value for this setting is too low
+        if ($Self->{DBObject}->{'DB::Type'} eq 'mysql') {
+
+            # check the actual setting
+            $Self->{DBObject}->Prepare(
+                SQL => "SHOW variables WHERE Variable_name = 'max_allowed_packet'",
+            );
+
+            my $MaxAllowedPacket            = 0;
+            my $MaxAllowedPacketRecommended = 20;
+            while ( my @Data = $Self->{DBObject}->FetchrowArray() ) {
+                if ($Data[1]) {
+                    $MaxAllowedPacket = $Data[1] / 1024 / 1024;
+                }
+            }
+
+            if ($MaxAllowedPacket < $MaxAllowedPacketRecommended) {
+                $Self->{LayoutObject}->Block(
+                    Name => 'DatabasePackageSizeWarning',
+                    Data => {
+                        MaxAllowedPacket            => $MaxAllowedPacket,
+                        MaxAllowedPacketRecommended => $MaxAllowedPacketRecommended,
+                    },
+                );
+            }
+        }
     }
 
     # FeatureAddons